+++ src/components/feedback-information/index.tsx
@@ -3,7 +3,7 @@ import { useContext } from "react";
 import Link from "next/link"
 import { ContextApplication } from "@/context";
 import { Button } from "@/components/button"
-import { ArrowIcon } from "../../../../../../public/assets/icons"
+import { ArrowIcon } from "../../../public/assets/icons"
 import { useEffect, useState } from "react"
