+++ src/components/topics/index.tsx
@@ -4,20 +4,20 @@ import {useEffect, useState} from "react"
 import Checkbox from "@/components/checkbox"
 import Details from "@/components/details"
 import Validation from "@/components/validation"
-import { descriptionDetail } from "../../../../../../util/description-detail"
-import { questions } from "../../../../../../util/questionsInfo"
+import { descriptionDetail } from "../../../util/description-detail"
+import { questions } from "../../../util/questionsInfo"
 import { useContext } from "react";
 import { ContextApplication } from "@/context";
 
 export const checkboxId:number[] = [3,4,5,6,7,8,9,10]
 
-const ImpactQuestion = () => {
+const TopicsQuestion = () => {
     const { onChangeQuestion, setQuestion, selectedCheckbox, setSelectedCheckbox } = useContext(ContextApplication);
     const [isError, setIsError] = useState<boolean>(false)
     const [idApplication, setId] = useState()
 
     useEffect(() => {
-        const getStorage = localStorage.getItem("impact") || '{}'
+        const getStorage = localStorage.getItem("topics") || '{}'
         const initialValue = JSON.parse(getStorage)
         const applicationStorage = localStorage.getItem("application") || '{}'
         const applicationIdStorage = JSON.parse(applicationStorage).id
@@ -29,8 +29,8 @@ const ImpactQuestion = () => {
     const onChecked = (e:any) => {
         const {id, checked} = e.target
         checked ? 
-            (setSelectedCheckbox([...selectedCheckbox, parseInt(id)]), localStorage.setItem('impact', JSON.stringify({id: idApplication, value: [...selectedCheckbox, parseInt(id)]}))): 
-            (setSelectedCheckbox([...selectedCheckbox?.filter(el => el !== parseInt(id))]), localStorage.setItem('impact', JSON.stringify({id: idApplication, value: [...selectedCheckbox?.filter(el => el !== parseInt(id))]})))
+            (setSelectedCheckbox([...selectedCheckbox, parseInt(id)]), localStorage.setItem('topics', JSON.stringify({id: idApplication, value: [...selectedCheckbox, parseInt(id)]}))): 
+            (setSelectedCheckbox([...selectedCheckbox?.filter(el => el !== parseInt(id))]), localStorage.setItem('topics', JSON.stringify({id: idApplication, value: [...selectedCheckbox?.filter(el => el !== parseInt(id))]})))
     }
 
     return(
@@ -38,7 +38,7 @@ const ImpactQuestion = () => {
             <BackLink content='Back'onClick={() => setQuestion(1)}/>
             <h1 className="govuk-heading-l">What topics do you want to comment on?</h1>
             <p className="govuk-body">Help us understand what your comments on this development is about. Select all the topics that apply.</p>
-            <Details summary="What happens to your comments" description={descriptionDetail['impact']}/>
+            <Details summary="What happens to your comments" description={descriptionDetail['topics']}/>
                 {
                     checkboxId.map(el => 
                         <Checkbox label={questions[el]} id={el.toString()} onChange={(e) => onChecked(e)} key={el} checked={selectedCheckbox.includes(el)}/>
@@ -47,9 +47,9 @@ const ImpactQuestion = () => {
         {
             isError && <Validation message="Please select at least one topic"/>
         }
-        <Button content="Next" className="button-impact-question" onClick={() => {selectedCheckbox.length > 0 ? onChangeQuestion() : setIsError(true)}}/>
+        <Button content="Next" className="button-topics-question" onClick={() => {selectedCheckbox.length > 0 ? onChangeQuestion() : setIsError(true)}}/>
         </section>
     )
 }
 
-export default ImpactQuestion;
+export default TopicsQuestion;
