+++ src/pages/planning-applications/[id]/feedback/index.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useContext } from "react"
 import { ContextApplication } from "@/context";
 import Breadcrumbs from "@/components/breadcrumbs"
 import Instructions from "./instructions"
-import Questions from "./questions"
+import Questions from "../../../../components/questions"
 
 
 const Feedback = () => {
