+++ src/styles/feedback.scss
@@ -60,8 +60,8 @@
     margin-top: 5px;
 }
 
-// impact style
-.button-impact-question {
+// topics style
+.button-topics-question {
     margin-top: 20px
 }
 .impact-validation {
